refactor(views): streamline ViewBuilder by integrating InputValidator, Aggregator, and Writer
- Removed obsolete methods for input checks and file validation from ViewBuilder.
- Introduced InputValidator for input data structure validation.
- Introduced Aggregator for metric aggregation (terms and temporal steps).
- Introduced Writer for consolidating results and metric structure parquet output.
- Updated tests for the new components and broader input validation scenarios.
