+++ gems_views_builder/common.py
@@ -1,0 +1,5 @@
+from typing import Literal
+
+PARQUET_COMPRESSION: Literal["zstd"] = "zstd"
+PARQUET_COMPRESSION_LEVEL = 3
+PARQUET_ROW_GROUP_SIZE = 64_000
