+++ tests/test_consolidate_results.py
@@ -18,9 +18,10 @@ import polars as pl
 import pyarrow.parquet as pq
 import pytest
 
-from gems_views_builder.views import _PARQUET_ROW_GROUP_SIZE, ViewBuilder
+from gems_views_builder.common import PARQUET_ROW_GROUP_SIZE
+from gems_views_builder.writer import Writer
 
-# Schema matching _aggregate_metric_temporally output
+# Schema matching temporal aggregation chunk output
 _CHUNK_SCHEMA = {
     "metric_id": pl.Utf8,
     "metric_location": pl.Utf8,
@@ -31,12 +32,6 @@ _CHUNK_SCHEMA = {
 }
 
 _RESULT_FILENAME_PATTERN = re.compile(r"^view\d{8}T\d{6}\.parquet$")
-
-
-def _make_builder(tmp_path: Path) -> ViewBuilder:
-    vb = object.__new__(ViewBuilder)
-    vb.input_data_path = tmp_path
-    return vb
 
 
 def _write_chunk(path: Path, rows: list[dict[str, object]]) -> Path:
@@ -95,46 +90,51 @@ def two_chunks(tmp_path: Path) -> list[Path]:
 # ---------------------------------------------------------------------------
 
 
+def test_consolidate_results_returns_none_and_writes_nothing_when_no_chunks(tmp_path: Path) -> None:
+    assert Writer(tmp_path).consolidate_results([]) is None
+    assert not (tmp_path / "results").exists()
+
+
 def test_result_file_is_created_in_results_dir(tmp_path: Path, two_chunks: list[Path]) -> None:
-    _make_builder(tmp_path)._consolidate_results(two_chunks)
+    Writer(tmp_path).consolidate_results(two_chunks)
     results = list((tmp_path / "results").glob("*.parquet"))
     assert len(results) == 1
 
 
 def test_result_filename_matches_pattern(tmp_path: Path, two_chunks: list[Path]) -> None:
-    _make_builder(tmp_path)._consolidate_results(two_chunks)
+    Writer(tmp_path).consolidate_results(two_chunks)
     result_file = next((tmp_path / "results").glob("*.parquet"))
     assert _RESULT_FILENAME_PATTERN.match(result_file.name), f"Unexpected filename: {result_file.name}"
 
 
 def test_results_dir_is_created_if_absent(tmp_path: Path, two_chunks: list[Path]) -> None:
     assert not (tmp_path / "results").exists()
-    _make_builder(tmp_path)._consolidate_results(two_chunks)
+    Writer(tmp_path).consolidate_results(two_chunks)
     assert (tmp_path / "results").is_dir()
 
 
 def test_chunk_files_are_deleted_after_consolidation(tmp_path: Path, two_chunks: list[Path]) -> None:
-    _make_builder(tmp_path)._consolidate_results(two_chunks)
+    Writer(tmp_path).consolidate_results(two_chunks)
     for chunk in two_chunks:
         assert not chunk.exists(), f"Chunk {chunk.name} was not deleted"
 
 
 def test_merged_file_contains_all_rows(tmp_path: Path, two_chunks: list[Path]) -> None:
-    _make_builder(tmp_path)._consolidate_results(two_chunks)
+    Writer(tmp_path).consolidate_results(two_chunks)
     result_file = next((tmp_path / "results").glob("*.parquet"))
     df = pl.read_parquet(result_file)
     assert len(df) == 3  # 2 rows from chunk_a + 1 row from chunk_b
 
 
 def test_merged_file_contains_correct_metric_ids(tmp_path: Path, two_chunks: list[Path]) -> None:
-    _make_builder(tmp_path)._consolidate_results(two_chunks)
+    Writer(tmp_path).consolidate_results(two_chunks)
     result_file = next((tmp_path / "results").glob("*.parquet"))
     metric_ids = set(pl.read_parquet(result_file)["metric_id"].to_list())
     assert metric_ids == {"a", "b"}
 
 
 def test_parquet_compression_is_zstd(tmp_path: Path, two_chunks: list[Path]) -> None:
-    _make_builder(tmp_path)._consolidate_results(two_chunks)
+    Writer(tmp_path).consolidate_results(two_chunks)
     result_file = next((tmp_path / "results").glob("*.parquet"))
     meta = cast(Any, pq).read_metadata(result_file)
     for rg_idx in range(meta.num_row_groups):
@@ -148,7 +148,7 @@ def test_parquet_row_group_size_respected(tmp_path: Path) -> None:
     """Write more rows than the row group size and verify multiple row groups are created."""
     chunks_dir = tmp_path / "temporal_aggregation"
     chunks_dir.mkdir()
-    n_rows = _PARQUET_ROW_GROUP_SIZE + 1
+    n_rows = PARQUET_ROW_GROUP_SIZE + 1
     chunk = _write_chunk(
         chunks_dir / "big-0.parquet",
         [
@@ -163,7 +163,7 @@ def test_parquet_row_group_size_respected(tmp_path: Path) -> None:
             for i in range(n_rows)
         ],
     )
-    _make_builder(tmp_path)._consolidate_results([chunk])
+    Writer(tmp_path).consolidate_results([chunk])
     result_file = next((tmp_path / "results").glob("*.parquet"))
     meta = cast(Any, pq).read_metadata(result_file)
     assert meta.num_row_groups >= 2, "Expected at least 2 row groups for data exceeding row_group_size"
@@ -185,7 +185,7 @@ def test_consolidate_with_single_chunk(tmp_path: Path) -> None:
             }
         ],
     )
-    _make_builder(tmp_path)._consolidate_results([chunk])
+    Writer(tmp_path).consolidate_results([chunk])
     result_file = next((tmp_path / "results").glob("*.parquet"))
     df = pl.read_parquet(result_file)
     assert len(df) == 1
