+++ tests/test_views_input_checks.py
@@ -10,64 +10,85 @@
 #
 # This file is part of the Antares project.
 
+"""Tests for ``InputValidator.validate()`` (required layout under ``input_data_path``)."""
+
+import shutil
 from pathlib import Path
 
 import pytest
 
-from gems_views_builder.views import EXACT_FILES, ViewBuilder
+from gems_views_builder.input_validator import EXACT_FILES, InputValidator
 
 
-def _builder_without_init(input_data_path: Path) -> ViewBuilder:
-    """ViewBuilder whose input checks can be invoked without loading YAML/parquet."""
-    vb = object.__new__(ViewBuilder)
-    vb.input_data_path = input_data_path
-    return vb
-
-
-def _touch_required_six_files(base: Path) -> None:
+def _write_minimal_valid_input_data(root: Path) -> None:
+    """Everything ``InputValidator.validate()`` expects: directory, catalogs, YAML set, calendar, simulation table."""
+    catalogs = root / "catalogs"
+    catalogs.mkdir(parents=True)
+    (catalogs / "placeholder.yml").touch()
     for name in EXACT_FILES:
-        (base / name).touch()
-    (base / "calendar.csv").touch()
-    (base / "simulation_table.parquet").touch()
+        (root / name).touch()
+    (root / "calendar.csv").touch()
+    (root / "simulation_table.parquet").touch()
 
 
-def test_check_required_input_files_passes_when_layout_valid(tmp_path: Path) -> None:
-    _touch_required_six_files(tmp_path)
-    _builder_without_init(tmp_path)._check_required_input_files()
+def test_validate_passes_for_minimal_valid_layout(tmp_path: Path) -> None:
+    _write_minimal_valid_input_data(tmp_path)
+    InputValidator(tmp_path).validate()
 
 
-def test_check_required_input_files_raises_when_exact_file_missing(tmp_path: Path) -> None:
-    _touch_required_six_files(tmp_path)
+def test_validate_raises_when_exact_file_missing(tmp_path: Path) -> None:
+    _write_minimal_valid_input_data(tmp_path)
     (tmp_path / "taxonomy.yml").unlink()
     with pytest.raises(FileNotFoundError, match="taxonomy.yml"):
-        _builder_without_init(tmp_path)._check_required_input_files()
+        InputValidator(tmp_path).validate()
 
 
-def test_check_required_input_files_raises_when_prefix_file_missing(tmp_path: Path) -> None:
-    _touch_required_six_files(tmp_path)
+def test_validate_raises_when_calendar_missing(tmp_path: Path) -> None:
+    _write_minimal_valid_input_data(tmp_path)
     (tmp_path / "calendar.csv").unlink()
     with pytest.raises(FileNotFoundError, match="calendar"):
-        _builder_without_init(tmp_path)._check_required_input_files()
+        InputValidator(tmp_path).validate()
 
 
-def test_check_required_input_files_raises_when_wrong_suffix_for_prefix_match(tmp_path: Path) -> None:
-    _touch_required_six_files(tmp_path)
+def test_validate_raises_when_calendar_has_wrong_suffix(tmp_path: Path) -> None:
+    _write_minimal_valid_input_data(tmp_path)
     (tmp_path / "calendar.csv").unlink()
     (tmp_path / "calendar.txt").touch()
     with pytest.raises(ValueError, match="must be a '.csv' file"):
-        _builder_without_init(tmp_path)._check_required_input_files()
+        InputValidator(tmp_path).validate()
 
 
-def test_check_required_input_files_raises_when_simulation_table_missing(tmp_path: Path) -> None:
-    _touch_required_six_files(tmp_path)
+def test_validate_raises_when_simulation_table_missing(tmp_path: Path) -> None:
+    _write_minimal_valid_input_data(tmp_path)
     (tmp_path / "simulation_table.parquet").unlink()
     with pytest.raises(FileNotFoundError, match="simulation_table"):
-        _builder_without_init(tmp_path)._check_required_input_files()
+        InputValidator(tmp_path).validate()
 
 
-def test_check_required_input_files_raises_when_wrong_suffix_for_simulation_table(tmp_path: Path) -> None:
-    _touch_required_six_files(tmp_path)
+def test_validate_raises_when_simulation_table_has_wrong_suffix(tmp_path: Path) -> None:
+    _write_minimal_valid_input_data(tmp_path)
     (tmp_path / "simulation_table.parquet").unlink()
     (tmp_path / "simulation_table.csv").touch()
     with pytest.raises(ValueError, match="must be a '.parquet' file"):
-        _builder_without_init(tmp_path)._check_required_input_files()
+        InputValidator(tmp_path).validate()
+
+
+def test_validate_raises_when_catalogs_directory_missing(tmp_path: Path) -> None:
+    _write_minimal_valid_input_data(tmp_path)
+    shutil.rmtree(tmp_path / "catalogs")
+    with pytest.raises(NotADirectoryError, match="catalogs"):
+        InputValidator(tmp_path).validate()
+
+
+def test_validate_raises_when_catalogs_directory_empty(tmp_path: Path) -> None:
+    _write_minimal_valid_input_data(tmp_path)
+    (tmp_path / "catalogs" / "placeholder.yml").unlink()
+    with pytest.raises(FileNotFoundError, match="empty"):
+        InputValidator(tmp_path).validate()
+
+
+def test_validate_raises_when_input_path_is_not_a_directory(tmp_path: Path) -> None:
+    file_path = tmp_path / "not_a_directory"
+    file_path.touch()
+    with pytest.raises(NotADirectoryError, match="not a directory"):
+        InputValidator(file_path).validate()
